+++ src/test/java/test/org/fugerit/java/nhg/ReflectDemo.java
@@ -1,0 +1,21 @@
+package test.org.fugerit.java.nhg;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class ReflectDemo {
+
+    private String field1;
+
+    public String getField1() {
+        return field1;
+    }
+
+    public void setField1(String field1) {
+        this.field1 = field1;
+    }
+
+    @Getter @Setter
+    private Integer field2;
+
+}